Removed some properties from getInitialState.
Apparently it's an anti-pattern
http://facebook.github.io/react/tips/props-in-getInitialState-as-anti-pattern.html

Thought removing isEditing from there was a bit harder though. It is
initial state, but maybe it's symptomatic of another problem. I should
be creating all components equally then dropping them into that state
later somehow.

I could have put it in the render method, but I thought that was really
weird.